refactor pact.io tests to attach interactions to the tests
