+++ tests/appsTest.js
@@ -12,6 +12,19 @@ fdescribe('Main: Currently testing apps management,', function () {
   var testApp = 'someAppName'
   var nonExistentApp = 'nonExistentApp' + timeRightNow
 
+  // PACT setup
+  const Pact = require('@pact-foundation/pact-web')
+  const provider = new Pact.PactWeb()
+  const { setGeneralInteractions } = require('./pactHelper.js')
+
+  beforeAll(function (done) {
+    Promise.all(setGeneralInteractions(provider)).then(done, done.fail)
+  })
+
+  afterAll(function (done) {
+    provider.removeInteractions().then(done, done.fail)
+  })
+
   beforeEach(function (done) {
     oc = new OwnCloud({
       baseUrl: config.owncloudURL,
