+++ tests/capabilitiesTest.js
@@ -5,6 +5,19 @@ fdescribe('Main: Currently testing getConfig, getVersion and getCapabilities', f
   // LIBRARY INSTANCE
   let oc
 
+  // PACT setup
+  const Pact = require('@pact-foundation/pact-web')
+  const provider = new Pact.PactWeb()
+  const { setGeneralInteractions } = require('./pactHelper.js')
+
+  beforeAll(function (done) {
+    Promise.all(setGeneralInteractions(provider)).then(done, done.fail)
+  })
+
+  afterAll(function (done) {
+    provider.removeInteractions().then(done, done.fail)
+  })
+
   beforeEach(function (done) {
     oc = new OwnCloud({
       baseUrl: config.owncloudURL,
