+++ tests/filesTest.js
@@ -9,6 +9,19 @@ fdescribe('Main: Currently testing files management,', function () {
   // LIBRARY INSTANCE
   let oc
 
+  // PACT setup
+  const Pact = require('@pact-foundation/pact-web')
+  const provider = new Pact.PactWeb()
+  const { setGeneralInteractions } = require('./pactHelper.js')
+
+  beforeAll(function (done) {
+    Promise.all(setGeneralInteractions(provider)).then(done, done.fail)
+  })
+
+  afterAll(function (done) {
+    provider.removeInteractions().then(done, done.fail)
+  })
+
   // TESTING CONFIGS
   const testContent = 'testContent'
   const testSubDir = config.testFolder + '/' + 'subdir'
