+++ tests/loginTest.js
@@ -11,6 +11,19 @@ fdescribe('Main: Currently testing Login and initLibrary,', function () {
   // TESTING CONFIGS
   var nonExistentUser = 'nonExistentUser' + timeRightNow
 
+  // PACT setup
+  const Pact = require('@pact-foundation/pact-web')
+  const provider = new Pact.PactWeb()
+  const { setGeneralInteractions } = require('./pactHelper.js')
+
+  beforeAll(function (done) {
+    Promise.all(setGeneralInteractions(provider)).then(done, done.fail)
+  })
+
+  afterAll(function (done) {
+    provider.removeInteractions().then(done, done.fail)
+  })
+
   beforeEach(function () {
     oc = null
   })
