+++ tests/shareRecipientTest.js
@@ -5,6 +5,19 @@ fdescribe('Main: Currently testing share recipient,', function () {
   // LIBRARY INSTANCE
   var oc
 
+  // PACT setup
+  const Pact = require('@pact-foundation/pact-web')
+  const provider = new Pact.PactWeb()
+  const { setGeneralInteractions } = require('./pactHelper.js')
+
+  beforeAll(function (done) {
+    Promise.all(setGeneralInteractions(provider)).then(done, done.fail)
+  })
+
+  afterAll(function (done) {
+    provider.removeInteractions().then(done, done.fail)
+  })
+
   // TESTING CONFIGS
   const testUserPassword = config.testUserPassword
   const testUser = config.testUser
