+++ tests/unauthorized/appsTest.js
@@ -7,7 +7,18 @@ fdescribe('Unauthorized: Currently testing apps management,', function () {
   // LIBRARY INSTANCE
   var oc
 
-  // TESTING CONFIGS
+  // PACT setup
+  const Pact = require('@pact-foundation/pact-web')
+  const provider = new Pact.PactWeb()
+  const { setGeneralInteractions } = require('../pactHelper.js')
+
+  beforeAll(function (done) {
+    Promise.all(setGeneralInteractions(provider)).then(done, done.fail)
+  })
+
+  afterAll(function (done) {
+    provider.removeInteractions().then(done, done.fail)
+  })
 
   beforeEach(function () {
     oc = new OwnCloud({
