+++ tests/unauthorized/capabilitiesTest.js
@@ -4,6 +4,19 @@ fdescribe('Unauthorized: Currently testing getConfig, getVersion and getCapabili
   // LIBRARY INSTANCE
   let oc
 
+  // PACT setup
+  const Pact = require('@pact-foundation/pact-web')
+  const provider = new Pact.PactWeb()
+  const { setGeneralInteractions } = require('../pactHelper.js')
+
+  beforeAll(function (done) {
+    Promise.all(setGeneralInteractions(provider)).then(done, done.fail)
+  })
+
+  afterAll(function (done) {
+    provider.removeInteractions().then(done, done.fail)
+  })
+
   beforeEach(function () {
     oc = new OwnCloud({
       baseUrl: config.owncloudURL,
