+++ tests/unauthorized/filesTest.js
@@ -7,6 +7,19 @@ describe('Unauthorized: Currently testing files management,', function () {
   // LIBRARY INSTANCE
   var oc
 
+  // PACT setup
+  const Pact = require('@pact-foundation/pact-web')
+  const provider = new Pact.PactWeb()
+  const { setGeneralInteractions } = require('../pactHelper.js')
+
+  beforeAll(function (done) {
+    Promise.all(setGeneralInteractions(provider)).then(done, done.fail)
+  })
+
+  afterAll(function (done) {
+    provider.removeInteractions().then(done, done.fail)
+  })
+
   // TESTING CONFIGS
   var testContent = 'testContent'
   var testFolder = '/testFolder' + timeRightNow
